Moving HcpFeaturedCard to own file
